StanfordXmlParser: finish naive implementation
The parser will parse specific XML files, map cat codes
to user-friendly cat text if possible, map non-duplicate
actors and films to the schema, and add non-duplicate
relationships.

For each new film added, a random price is assigned between
$1.00 and $100.00 (same as `sample_prices.sql`)

Casts are checked for foreign key constraints. For sake
of simplicity, any actors of the same stage name are
associated with the relationship. This is reported as a
warning if there are multiple but remains schema consistent.

The parser also avoids potential id conflicts by prepending
a unique header ("sa" for actors, "sf" for films). As an example,
`fid='TTz9'` would conflict with the SQL stored procedure.

Note the runtime of the naive implementation as tested on my
machine is around 480s-660s (8-11 minutes).